Moved trigger attributes, separated attributes into different files.  Added supply trigger. MSF-33
